♻️ 다시 한번 리팩토링
