+++ src/app/request/_components/LocationModal.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import React from 'react';
+import { Continent } from '../_types/form';
 
 type Props = {
   isOpen: boolean;
@@ -8,7 +9,7 @@ type Props = {
   handleLocationSelect: (location: string) => void;
 };
 
-const continents = [
+const continents:Continent[] = [
   { name: '아시아', cities: ['서울', '도쿄', '방콕', '싱가포르'] },
   { name: '유럽', cities: ['런던', '파리', '로마', '베를린'] },
   { name: '북미', cities: ['뉴욕', '로스앤젤레스', '토론토', '시애틀'] },
