+++ src/app/request/_types/form.ts
@@ -1,0 +1,13 @@
+export type FormInputs = {
+  location: string;
+  topic: string;
+  title: string;
+  credit: number;
+  content: string;
+  deadline: string;
+};
+
+export type Continent = {
+  name: string;
+  cities: string[];
+};
