--- src/app/request/types/form.ts
@@ -1,8 +1,0 @@
-export type FormInputs = {
-    location: string;
-    topic: string;
-    title: string;
-    credit: number;
-    content: string;
-    deadline: string;
-  };
